@@ -4,15 +4,15 @@ version: 0.10.0
 homepage: https://github.com/blackmann/locationpicker
 
 environment:
-  sdk: ">=2.14.0 <3.0.0"
+  sdk: '>=3.3.2 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.5
-  google_maps_flutter: ^2.2.0
+  http: any
+  google_maps_flutter: any
 #  location: ^4.4.0
-  geolocator: ^9.0.1
+  geolocator: any
 
 dev_dependencies:
   flutter_test:
